ISSUE #2865 regenerateAuthSession and endSession moved in auth output middleware userAgent helper created
