+++ backend/src/v5/models/loginRecord.js
@@ -16,11 +16,12 @@
  */
 
 // detects edge as browser but not device
-const { getLocationFromIPAddress, getUserAgentInfoFromBrowser, getUserAgentInfoFromPlugin,
-	isUserAgentFromPlugin } = require('../services/sessions');
+const { getUserAgentInfoFromBrowser, getUserAgentInfoFromPlugin, 
+	isUserAgentFromPlugin } = require('../utils/helper/userAgent');
 const db = require('../handler/db');
 const { events } = require('../services/eventsManager/eventsManager.constants');
 const { publish } = require('../services/eventsManager/eventsManager');
+const geoip = require('geoip-lite');
 
 const LoginRecord = {};
 
@@ -35,7 +36,7 @@ LoginRecord.saveLoginRecord = async (username, sessionId, ipAddress, userAgent, 
 		...uaInfo 
 	};
 
-	const location = getLocationFromIPAddress(loginRecord.ipAddr);
+	const location = geoip.lookup(ipAddress);
 	loginRecord.location = {
 		country: location?.country ?? 'unknown',
 		city: location?.city ?? 'unknown',
